+++ app/src/main/java/com/fesskiev/mediacenter/data/model/MediaFile.java
@@ -13,7 +13,7 @@ public interface MediaFile {
 
     String getArtworkPath();
 
-    int getLength();
+    long getLength();
 
     long getSize();
